Refactor add edit note package name from add_edit_note to addeditnote
