+++ app/src/main/java/com/dogactnrvrdi/notesapp/presentation/MainActivity.kt
@@ -4,18 +4,13 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
-import androidx.compose.ui.Modifier
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
-import com.dogactnrvrdi.notesapp.presentation.add_edit_note.AddEditNoteScreen
+import com.dogactnrvrdi.notesapp.presentation.addeditnote.AddEditNoteScreen
 import com.dogactnrvrdi.notesapp.presentation.notes.NotesScreen
 import com.dogactnrvrdi.notesapp.presentation.theme.NotesAppTheme
 import dagger.hilt.android.AndroidEntryPoint
