+++ app/src/main/java/com/dogactnrvrdi/notesapp/presentation/addeditnote/components/DescriptionTextField.kt
@@ -1,58 +1,54 @@
-package com.dogactnrvrdi.notesapp.presentation.add_edit_note.components
+package com.dogactnrvrdi.notesapp.presentation.addeditnote.components
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.text.BasicTextField
 import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material3.Text
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.TextField
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.focus.FocusState
-import androidx.compose.ui.focus.onFocusChanged
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.input.KeyboardCapitalization
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import com.dogactnrvrdi.notesapp.R
 
 @Composable
-fun TitleTextField(
+fun DescriptionTextField(
     text: String,
     modifier: Modifier = Modifier,
-    onValueChange: (String) -> Unit
+    onValueChange: (String) -> Unit,
 ) {
     Box(modifier = modifier) {
 
         OutlinedTextField(
             value = text,
             onValueChange = { newValue ->
-                onValueChange(capitalizeEachWord(newValue))
+                onValueChange(capitalizeFirstLetter(newValue))
             },
-            label = { Text(text = stringResource(R.string.title)) },
-            maxLines = 3,
             textStyle = TextStyle(
                 color = MaterialTheme.colorScheme.onBackground,
-                fontSize = MaterialTheme.typography.headlineMedium.fontSize
+                fontSize = MaterialTheme.typography.titleLarge.fontSize
             ),
+            label = { Text(text = stringResource(R.string.description)) },
             keyboardOptions = KeyboardOptions.Default.copy(
-                capitalization = KeyboardCapitalization.Words
+                capitalization = KeyboardCapitalization.Sentences
             ),
             modifier = Modifier
                 .fillMaxWidth()
-                .padding(horizontal = 10.dp)
+                .fillMaxHeight()
+                .padding(start = 10.dp)
         )
     }
 }
 
-fun capitalizeEachWord(input: String): String {
-    return input.split(" ")
-        .joinToString(" ") { it.replaceFirstChar { char -> char.uppercaseChar() } }
+fun capitalizeFirstLetter(input: String): String {
+    return if (input.isNotEmpty()) {
+        input.replaceFirstChar { it.uppercaseChar() }
+    } else {
+        input
+    }
 }
